backend/requirements.txt
Django>=5.0,<6.0
django-allauth[socialaccount]>=65.0,<66.0
python-decouple>=3.8
whitenoise>=6.7
